# faster intersects check