--- a/geo/benches/intersection.rs
+++ b/geo/benches/intersection.rs
@@ -7,11 +7,11 @@ use criterion::Criterion;
 use geo::intersects::Intersects;
 use geo::MultiPolygon;
 
-fn criterion_benchmark(c: &mut Criterion) {
+fn multi_polygon_intersection(c: &mut Criterion) {
     let plot_polygons: MultiPolygon = geo_test_fixtures::nl_plots();
     let zone_polygons: MultiPolygon = geo_test_fixtures::nl_zones();
 
-    c.bench_function("intersection", |bencher| {
+    c.bench_function("MultiPolygon intersects", |bencher| {
         bencher.iter(|| {
             let mut intersects = 0;
             let mut non_intersects = 0;
@@ -32,10 +32,44 @@ fn criterion_benchmark(c: &mut Criterion) {
     });
 }
 
-criterion_group! {
-    name = benches;
-    config = Criterion::default().sample_size(10);
-    targets = criterion_benchmark
+fn rect_intersection(c: &mut Criterion) {
+    use geo::algorithm::BoundingRect;
+    use geo::geometry::Rect;
+    let plot_bbox: Vec<Rect> = geo_test_fixtures::nl_plots()
+        .iter()
+        .map(|plot| plot.bounding_rect().unwrap())
+        .collect();
+    let zone_bbox: Vec<Rect> = geo_test_fixtures::nl_zones()
+        .iter()
+        .map(|plot| plot.bounding_rect().unwrap())
+        .collect();
+
+    c.bench_function("Rect intersects", |bencher| {
+        bencher.iter(|| {
+            let mut intersects = 0;
+            let mut non_intersects = 0;
+
+            for a in &plot_bbox {
+                for b in &zone_bbox {
+                    if criterion::black_box(a.intersects(b)) {
+                        intersects += 1;
+                    } else {
+                        non_intersects += 1;
+                    }
+                }
+            }
+
+            assert_eq!(intersects, 3054);
+            assert_eq!(non_intersects, 25702);
+        });
+    });
 }
 
-criterion_main!(benches);
+criterion_group! {
+    name = bench_multi_polygons;
+    config = Criterion::default().sample_size(10);
+    targets = multi_polygon_intersection
+}
+criterion_group!(bench_rects, rect_intersection);
+
+criterion_main!(bench_multi_polygons, bench_rects);