- [x] I agree to follow the project's [code of conduct](https://github.com/georust/geo/blob/main/CODE_OF_CONDUCT.md).
- [x] I added an entry to `CHANGES.md` if knowledge of this change could be valuable to users.
---
    
The code change is only for the Rect intersection check, but other geometry types, like LineString and Polygon, leverage Rect::intersects, so we should see improvements in almost all of our disjoint intersection checks.
    
cargo bench excerpt:
    
    MultiPolygon intersects time:   [476.96 ms 477.13 ms 477.32 ms]
                            change: [-8.4983% -8.3486% -8.2284%] (p = 0.00 < 0.05)
                            Performance has improved.
    Rect intersects         time:   [27.498 µs 27.546 µs 27.598 µs]
                            change: [-57.503% -57.444% -57.389%] (p = 0.00 < 0.05)
                            Performance has improved.